+++ rvs/runtime/java.py
@@ -16,7 +16,7 @@ from pathlib import Path
 from typing import Any
 from urllib.parse import urlsplit
 
-import httpx
+import httpx2 as httpx
 
 from .. import output
 from ._install import (